+++ ios/RN/RNCamera.m
@@ -21,6 +21,7 @@
 @property (nonatomic, copy) RCTDirectEventBlock onMountError;
 @property (nonatomic, copy) RCTDirectEventBlock onBarCodeRead;
 @property (nonatomic, copy) RCTDirectEventBlock onFacesDetected;
+@property (nonatomic, copy) RCTDirectEventBlock onPictureSaved;
 
 @end
 
@@ -82,6 +83,13 @@ static NSDictionary *defaultFaceDetectorOptions = nil;
 {
     if (_onBarCodeRead) {
         _onBarCodeRead(event);
+    }
+}
+
+- (void)onPictureSaved:(NSDictionary *)event
+{
+    if (_onPictureSaved) {
+        _onPictureSaved(event);
     }
 }
 
@@ -210,7 +218,7 @@ static NSDictionary *defaultFaceDetectorOptions = nil;
     AVCaptureDevice *device = [self.videoCaptureDeviceInput device];
     NSError *error = nil;
 
-    if (self.autoFocus < 0 || device.focusMode != RNCameraAutoFocusOff || device.position == RNCameraTypeFront) {
+    if (device == nil || self.autoFocus < 0 || device.focusMode != RNCameraAutoFocusOff || device.position == RNCameraTypeFront) {
         return;
     }
 
@@ -284,6 +292,11 @@ static NSDictionary *defaultFaceDetectorOptions = nil;
     [device unlockForConfiguration];
 }
 
+- (void)updatePictureSize
+{
+    [self updateSessionPreset:self.pictureSize];
+}
+
 #if __has_include(<GoogleMobileVision/GoogleMobileVision.h>)
 - (void)updateFaceDetecting:(id)faceDetecting
 {
@@ -309,19 +322,33 @@ static NSDictionary *defaultFaceDetectorOptions = nil;
 - (void)takePicture:(NSDictionary *)options resolve:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject
 {
     AVCaptureConnection *connection = [self.stillImageOutput connectionWithMediaType:AVMediaTypeVideo];
-    [connection setVideoOrientation:[RNCameraUtils videoOrientationForDeviceOrientation:[[UIDevice currentDevice] orientation]]];
+    int orientation;
+    if ([options[@"orientation"] integerValue]) {
+        orientation = [options[@"orientation"] integerValue];
+    } else {
+        orientation = [RNCameraUtils videoOrientationForDeviceOrientation:[[UIDevice currentDevice] orientation]];
+    }
+    [connection setVideoOrientation:orientation];
     [self.stillImageOutput captureStillImageAsynchronouslyFromConnection:connection completionHandler: ^(CMSampleBufferRef imageSampleBuffer, NSError *error) {
         if (imageSampleBuffer && !error) {
+            BOOL useFastMode = options[@"fastMode"] && [options[@"fastMode"] boolValue];
+            if (useFastMode) {
+                resolve(nil);
+            }
             NSData *imageData = [AVCaptureStillImageOutput jpegStillImageNSDataRepresentation:imageSampleBuffer];
 
             UIImage *takenImage = [UIImage imageWithData:imageData];
 
-            CGRect frame = [_previewLayer metadataOutputRectOfInterestForRect:self.frame];
             CGImageRef takenCGImage = takenImage.CGImage;
-            size_t width = CGImageGetWidth(takenCGImage);
-            size_t height = CGImageGetHeight(takenCGImage);
-            CGRect cropRect = CGRectMake(frame.origin.x * width, frame.origin.y * height, frame.size.width * width, frame.size.height * height);
-            takenImage = [RNImageUtils cropImage:takenImage toRect:cropRect];
+            CGSize previewSize;
+            if (UIInterfaceOrientationIsPortrait([[UIApplication sharedApplication] statusBarOrientation])) {
+                previewSize = CGSizeMake(self.previewLayer.frame.size.height, self.previewLayer.frame.size.width);
+            } else {
+                previewSize = CGSizeMake(self.previewLayer.frame.size.width, self.previewLayer.frame.size.height);
+            }
+            CGRect cropRect = CGRectMake(0, 0, CGImageGetWidth(takenCGImage), CGImageGetHeight(takenCGImage));
+            CGRect croppedSize = AVMakeRectWithAspectRatioInsideRect(previewSize, cropRect);
+            takenImage = [RNImageUtils cropImage:takenImage toRect:croppedSize];
 
             if ([options[@"mirrorImage"] boolValue]) {
                 takenImage = [RNImageUtils mirrorImage:takenImage];
@@ -371,7 +398,11 @@ static NSDictionary *defaultFaceDetectorOptions = nil;
                 [RNImageUtils updatePhotoMetadata:imageSampleBuffer withAdditionalData:@{ @"Orientation": @(imageRotation) } inResponse:response]; // TODO
             }
 
-            resolve(response);
+            if (useFastMode) {
+                [self onPictureSaved:@{@"data": response, @"id": options[@"id"]}];
+            } else {
+                resolve(response);
+            }
         } else {
             reject(@"E_IMAGE_CAPTURE_FAILED", @"Image could not be captured", error);
         }
@@ -410,18 +441,21 @@ static NSDictionary *defaultFaceDetectorOptions = nil;
     [self updateSessionAudioIsMuted:!!options[@"mute"]];
 
     AVCaptureConnection *connection = [self.movieFileOutput connectionWithMediaType:AVMediaTypeVideo];
-    [connection setVideoOrientation:[RNCameraUtils videoOrientationForInterfaceOrientation:[[UIApplication sharedApplication] statusBarOrientation]]];
-    
-    
+    if (self.videoStabilizationMode != 0) {
+        if (connection.isVideoStabilizationSupported == NO) {
+            RCTLogWarn(@"%s: Video Stabilization is not supported on this device.", __func__);
+        } else {
+            [connection setPreferredVideoStabilizationMode:self.videoStabilizationMode];
+        }
+    }
+    [connection setVideoOrientation:[RNCameraUtils videoOrientationForDeviceOrientation:[[UIDevice currentDevice] orientation]]];
+
     if (options[@"codec"]) {
+      if (@available(iOS 10, *)) {
         AVVideoCodecType videoCodecType = options[@"codec"];
-        if (@available(iOS 10, *)) {
-            if ([self.movieFileOutput.availableVideoCodecTypes containsObject:videoCodecType]) {
-                [self.movieFileOutput setOutputSettings:@{AVVideoCodecKey:videoCodecType} forConnection:connection];
-                self.videoCodecType = videoCodecType;
-            } else {
-                RCTLogWarn(@"%s: Video Codec '%@' is not supported on this device.", __func__, videoCodecType);
-            }
+        if ([self.movieFileOutput.availableVideoCodecTypes containsObject:videoCodecType]) {
+          [self.movieFileOutput setOutputSettings:@{AVVideoCodecKey:videoCodecType} forConnection:connection];
+          self.videoCodecType = videoCodecType;
         } else {
             RCTLogWarn(@"%s: Setting videoCodec is only supported above iOS version 10.", __func__);
         }
@@ -429,15 +463,21 @@ static NSDictionary *defaultFaceDetectorOptions = nil;
     
     dispatch_async(self.sessionQueue, ^{
         [self updateFlashMode];
-        NSString *path = [RNFileSystem generatePathInDirectory:[[RNFileSystem cacheDirectoryPath] stringByAppendingString:@"Camera"] withExtension:@".mov"];
-        
+        NSString *path = nil;
+        if (options[@"path"]) {
+            path = options[@"path"];
+        }
+        else {
+            path = [RNFileSystem generatePathInDirectory:[[RNFileSystem cacheDirectoryPath] stringByAppendingPathComponent:@"Camera"] withExtension:@".mov"];
+        }
+
         if ([options[@"mirrorVideo"] boolValue]) {
             if ([connection isVideoMirroringSupported]) {
                 [connection setAutomaticallyAdjustsVideoMirroring:NO];
                 [connection setVideoMirrored:YES];
             }
         }
-        
+
         NSURL *outputURL = [[NSURL alloc] initFileURLWithPath:path];
         [self.movieFileOutput startRecordingToOutputFileURL:outputURL recordingDelegate:self];
         self.videoRecordedResolve = resolve;
@@ -448,6 +488,16 @@ static NSDictionary *defaultFaceDetectorOptions = nil;
 - (void)stopRecording
 {
     [self.movieFileOutput stopRecording];
+}
+
+- (void)resumePreview
+{
+    [[self.previewLayer connection] setEnabled:YES];
+}
+
+- (void)pausePreview
+{
+    [[self.previewLayer connection] setEnabled:NO];
 }
 
 - (void)startSession
@@ -465,6 +515,8 @@ static NSDictionary *defaultFaceDetectorOptions = nil;
             return;
         }
 
+        self.session.sessionPreset = AVCaptureSessionPresetPhoto;
+        
         AVCaptureStillImageOutput *stillImageOutput = [[AVCaptureStillImageOutput alloc] init];
         if ([self.session canAddOutput:stillImageOutput]) {
             stillImageOutput.outputSettings = @{AVVideoCodecKey : AVVideoCodecJPEG};
@@ -572,10 +624,17 @@ static NSDictionary *defaultFaceDetectorOptions = nil;
 
 #pragma mark - internal
 
-- (void)updateSessionPreset:(NSString *)preset
+- (void)updateSessionPreset:(AVCaptureSessionPreset)preset
 {
 #if !(TARGET_IPHONE_SIMULATOR)
+    if ([preset integerValue] < 0) {
+        return;
+    }
     if (preset) {
+        if (self.isDetectingFaces && [preset isEqual:AVCaptureSessionPresetPhoto]) {
+            RCTLog(@"AVCaptureSessionPresetPhoto not supported during face detection. Falling back to AVCaptureSessionPresetHigh");
+            preset = AVCaptureSessionPresetHigh;
+        }
         dispatch_async(self.sessionQueue, ^{
             [self.session beginConfiguration];
             if ([self.session canSetSessionPreset:preset]) {
@@ -764,19 +823,21 @@ static NSDictionary *defaultFaceDetectorOptions = nil;
         }
     }
     if (success && self.videoRecordedResolve != nil) {
-        AVVideoCodecType videoCodec = self.videoCodecType;
-        if (videoCodec == nil) {
-            videoCodec = [self.movieFileOutput.availableVideoCodecTypes firstObject];
-        }
-        if ([connections[0] isVideoMirrored]) {
+      if (@available(iOS 10, *)) {
+          AVVideoCodecType videoCodec = self.videoCodecType;
+          if (videoCodec == nil) {
+              videoCodec = [self.movieFileOutput.availableVideoCodecTypes firstObject];
+          }
+          if ([connections[0] isVideoMirrored]) {
             [self mirrorVideo:outputFileURL completion:^(NSURL *mirroredURL) {
                 self.videoRecordedResolve(@{ @"uri": mirroredURL.absoluteString, @"codec":videoCodec });
-                [self cleanupCamera];
             }];
-        } else {
+          } else {
             self.videoRecordedResolve(@{ @"uri": outputFileURL.absoluteString, @"codec":videoCodec });
-            [self cleanupCamera];
-        }
+          }
+      } else {
+          self.videoRecordedResolve(@{ @"uri": outputFileURL.absoluteString });
+      }
     } else if (self.videoRecordedReject != nil) {
         self.videoRecordedReject(@"E_RECORDING_FAILED", @"An error occurred while recording a video.", error);
     }
@@ -796,8 +857,8 @@ static NSDictionary *defaultFaceDetectorOptions = nil;
     [_faceDetectorManager maybeStartFaceDetectionOnSession:_session withPreviewLayer:_previewLayer];
 #endif
 
-    if (self.session.sessionPreset != AVCaptureSessionPresetHigh) {
-        [self updateSessionPreset:AVCaptureSessionPresetHigh];
+    if (self.session.sessionPreset != AVCaptureSessionPresetPhoto) {
+        [self updateSessionPreset:AVCaptureSessionPresetPhoto];
     }
 }