+++ src/handlePermissions.js
@@ -19,8 +19,7 @@ export const requestPermissions = async (hasVideoAndAudio, CameraManager, permis
           // On devices before SDK version 23, the permissions are automatically granted if they appear in the manifest,
           // so check and request should always be true.
           // https://github.com/facebook/react-native-website/blob/master/docs/permissionsandroid.md
-          const isAuthorized =
-            Platform.Version >= 23 ? granted === PermissionsAndroid.RESULTS.GRANTED : granted === true;
+          const isAuthorized = granted === PermissionsAndroid.RESULTS.GRANTED || granted === true;
     
           return isAuthorized;
     }