+++ types/index.d.ts
@@ -9,24 +9,39 @@
  *
  * If you are seeing this from the future, please, send us your cutting-edge technology :) (if it exists)
  */
-import { Component } from 'react';
+import { Component, ReactNode } from 'react';
 import { ViewProperties } from "react-native";
 
-type AutoFocus = { on: any, off: any };
-type FlashMode = { on: any, off: any, torch: any, auto: any };
-type CameraType = { front: any, back: any };
-type WhiteBalance = { sunny: any, cloudy: any, shadow: any, incandescent: any, fluorescent: any, auto: any };
-type BarCodeType = { aztec: any, code128: any, code39: any, code39mod43: any, code93: any, ean13: any, ean8: any, pdf417: any, qr: any, upce: any, interleaved2of5: any, itf14: any, datamatrix: any };
-type VideoQuality = { '2160p': any, '1080p': any, '720p': any, '480p': any, '4:3': any };
-type VideoCodec = { 'H264': symbol, 'JPEG': symbol, 'HVEC': symbol, 'AppleProRes422': symbol, 'AppleProRes4444': symbol };
+type AutoFocus = Readonly<{ on: any, off: any }>;
+type FlashMode = Readonly<{ on: any, off: any, torch: any, auto: any }>;
+type CameraType = Readonly<{ front: any, back: any }>;
+type WhiteBalance = Readonly<{ sunny: any, cloudy: any, shadow: any, incandescent: any, fluorescent: any, auto: any }>;
+type BarCodeType = Readonly<{ aztec: any, code128: any, code39: any, code39mod43: any, code93: any, ean13: any, ean8: any, pdf417: any, qr: any, upce: any, interleaved2of5: any, itf14: any, datamatrix: any }>;
+type VideoQuality = Readonly<{
+    '2160p': any, '1080p': any, '720p': any, '480p': any, '4:3': any;
+    /** iOS Only. Android not supported. */
+    '288p': any;
+}>;
+type VideoCodec = Readonly<{ 'H264': symbol, 'JPEG': symbol, 'HVEC': symbol, 'AppleProRes422': symbol, 'AppleProRes4444': symbol }>;
 
-type FaceDetectionClassifications = { all: any, none: any };
-type FaceDetectionLandmarks = { all: any, none: any };
-type FaceDetectionMode = { fast: any, accurate: any };
-type GoogleVisionBarcodeType = { CODE_128: any, CODE_39: any, CODABAR: any, DATA_MATRIX: any, EAN_13: any, EAN_8: any, ITF: any,
-    QR_CODE: any, UPC_A: any, UPC_E: any, PDF417: any, AZTEC: any }
+type FaceDetectionClassifications = Readonly<{ all: any, none: any }>;
+type FaceDetectionLandmarks = Readonly<{ all: any, none: any }>;
+type FaceDetectionMode = Readonly<{ fast: any, accurate: any }>;
+type GoogleVisionBarcodeType = Readonly<{
+    CODE_128: any, CODE_39: any, CODABAR: any, DATA_MATRIX: any, EAN_13: any, EAN_8: any,
+    ITF: any, QR_CODE: any, UPC_A: any, UPC_E: any, PDF417: any, AZTEC: any
+}>;
+
+// FaCC (Function as Child Components)
+type Self<T> = { [P in keyof T]: P }
+type CameraStatus = Readonly<Self<{ READY: any, PENDING_AUTHORIZATION: any, NOT_AUTHORIZED: any }>>;
+type FaCC = (params: {
+    camera: RNCamera,
+    status: keyof CameraStatus
+}) => JSX.Element;
 
 export interface Constants {
+    CameraStatus: CameraStatus;
     AutoFocus: AutoFocus;
     FlashMode: FlashMode;
     VideoCodec: VideoCodec;
@@ -45,6 +60,8 @@ export interface Constants {
 }
 
 export interface RNCameraProps {
+    children?: ReactNode | FaCC;
+
     autoFocus?: keyof AutoFocus;
     type?: keyof CameraType;
     flashMode?: keyof FlashMode;